requirements.txt
Flask==3.0.0
Flask_Cors==4.0.0
moviepy==1.0.3
Pillow==10.1.0
pytesseract==0.3.10
Requests==2.31.0
SpeechRecognition==3.10.0